rfac: Update /spec folder structure
